fix(tests): sqlite migrations and use file sqlite
Check the driver name of the database connection rather than use
the environment variable DB_DATABASE to not need DB_DATABASE set.

Change multiple calls of dropColumn to one with an array in order to support
SQLite.

Create dedicated database connection for testing in order to hardcode
the database path, which allows us to not require a .env.testing for
tests to work.

Change the SQLite database to be file-based, as that (on my system) was
faster than the in-memory database, but YMMV.