Optimized and fixed position queries
This also fixes #566 as it didn't find position due to case sensitive. What I fixed in this issue was to remove the uneccesarry all() function that get all positions in database, rather went straight to firstWhere(). This function is case sensitive when done in a Illuminate\Collection, but not in sql query builder. Hence it's behaviour changed when I removed all()