ci: add & configure release-please (#586)
Adds release-please by configuring the release-please GitHub Action.

The action provides a continually up-to-date PR that can be merged when the main branch is ready to be released as a new version. Through conventional commit messages, it'll determine what the following version number is.

The action targets the composer project alone.