docs: fix contributors, update plugins & push on docs/* (#754)
Fixes the contributors plugin by updating to v2.2.3 after submitting my
own PR to fix a repo path issue. Adjusts the branch to point to main.

Removes the old configuration that we haven't gotten around to yet and
adds a few comments to help explain the plugins.

Allows all docs/* branches to push to the development documentation.