fix: being able to submit multiple votes
There was a missing policy check when casting the vote. So the script managed to add the vote to the alternative, but then failed when trying to register the user as casted vote (MYSQL dupe error).

This commit fixes it and enforces a better policy when casting a vote not making this possible anymore.