fix: content loss warning when editing reports
Not really a fix, but the functionality is rarely used, so it's most time consuming to try avoid users to do this mistake.

You now get a warning about content loss before you upload files to a report edit. Prior it was no way to know before it was too late.

Fixes #913