fix(training): persist one-time links for longer
Fixes #1287. 

Introduces regression tests as well as the solution to the problem. Note
that the specific bug fixed is more limited: if you browse to other
pages, the session one-time key will no longer be valid. This means that
you can't lookup other sessions, dates, or even open up the dashboard
without losing the one-time key which is necessary for the redirect to
work.

This fix also limits the ability to *store* such a report in the event
the one-time key is invalid, whereas before it would partially work even
though it shouldn't have.

tests: adds several one-time link tests
  Adds a relatively comprehensive suite of tests to ensure it works as
  expected.

tests: re-enables the examination store test by faking notifications
  By faking notifications, we can easily re-enable these tests.

refactor: isExaminer()
  Includes a minor refactor of isExaminer() that aims to simplify the
  method.

Release-As: 6.3.5